microtubules/requirements.txt
magicgui==0.6.1
napari==0.4.17
numpy==1.20.1
opencv_python==4.6.0.66
Pillow==9.3.0
scikit_image==0.18.1
scipy==1.6.2
skimage==0.0
